+++ extensions/cursor-recent-projects/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Cursor Search Recent Projects Changelog
 
+## [Features] - 2024-10-11
+
+- Add integration with [Cursor Directory](https://www.raycast.com/escwxyz/cursor-directory)
+
 ## [Initial Version] - 2024-09-25
 
 - Add similar [VSCode](https://www.raycast.com/thomas/visual-studio-code) commands for Cursor